Creating Phoenix JDBC sample
