--- jdbcsamples/PhoenixJBDCSample/TestPhoenixReadMe.md
@@ -1,8 +1,0 @@
-**Overview**
-- This is just a simple jdbc program to show how the phoenix endpoints work.
-- To run this sample, you need to have an IAE Cluster up and running. 
-- Note down the endpoints from the VCAP of your service instance.
-- You also need a phoenix 4.9 client Java libraries
-- For more details refer to : https://console.bluemix.net/docs/services/AnalyticsEngine/working-with-hbase.html#working-with-hbase
-
-
