+++ jdbcsamples/TestPhoenix/.classpath
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="src" path="src"/>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/CDC-1.1%Foundation-1.1"/>
+	<classpathentry kind="lib" path="lib/phoenix-4.9.0-HBase-1.1-client.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-4.9.0-HBase-1.1-hive.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-4.9.0-HBase-1.1-pig.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-4.9.0-HBase-1.1-queryserver.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-4.9.0-HBase-1.1-server.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-4.9.0-HBase-1.1-thin-client.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-core-4.9.0-HBase-1.1.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-flume-4.9.0-HBase-1.1.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-hive-4.9.0-HBase-1.1.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-pherf-4.9.0-HBase-1.1-minimal.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-pherf-4.9.0-HBase-1.1.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-pig-4.9.0-HBase-1.1.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-queryserver-4.9.0-HBase-1.1.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-queryserver-client-4.9.0-HBase-1.1.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-spark-4.9.0-HBase-1.1.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-tracing-webapp-4.9.0-HBase-1.1-runnable.jar"/>
+	<classpathentry kind="lib" path="lib/phoenix-tracing-webapp-4.9.0-HBase-1.1.jar"/>
+	<classpathentry kind="output" path="bin"/>
+</classpath>
